Buffer and docs improvements
- Enable all features for building and publishing docs
- Improve and add some doc comments and examples
- Move Parser buffer implementations to separate buffer module
- Add a `FixedBuffer` implementation that owns the byte array
- Add ParserBuilder to simplify the instantiation of the Parser
